tools/pyproject.toml
[tool.poetry]
name = "i9000-tools"
version = "0.1.0"
description = ""
authors = ["Devin McCabe <devin.mccabe@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Unidecode = "^1.2.0"
boto3 = "^1.17.39"
python3-discogs-client = "^2.3.12"
python-dotenv = "^0.19.1"
psycopg2 = "^2.9.3"
musicbrainzngs = "^0.7.1"
pandas = "^2.2.0"

[tool.poetry.dev-dependencies]
black = "^22.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
